+++ packages/circular-progress-bar/src/types/component-size.ts
@@ -1,1 +1,1 @@
-export type ComponentSize = 'xxl' | 'xl' | 'l' | 'm' | 's' | 'xs' | 24 | 48 | 64 | 80 | 128 | 144;
+export type ComponentSize = 24 | 48 | 64 | 80 | 128 | 144;