+++ packages/picker-button/src/mobile/Component.mobile.tsx
@@ -73,7 +73,7 @@ export const PickerButtonMobile = forwardRef<HTMLInputElement, PickerButtonMobil
                 }}
                 Field={Field}
                 Optgroup={Optgroup}
-                size={size === 'm' || size === 56 ? 56 : 48}
+                size={size === 56 ? 56 : 48}
                 closeOnSelect={true}
                 fieldProps={{
                     ...fieldDefaultProps,