+++ packages/universal-modal/src/mobile/components/buttons/arrow-button/index.ts
@@ -1,1 +1,0 @@
-export { ArrowButtonMobile } from './arrow-button';